pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leo-gym"               
version = "0.1.0"
description = "Low Earth Orbit Gym: A library for training RL-policies for control of satellites operating in LEO"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Nektarios Aristeidis Tafanidis", email = "Nektariostaf@gmail.com" }]
license = { text = "MIT" }

dependencies = [
  "cvxpy",
  "gymnasium==0.28.1",
  "matplotlib==3.8.2",
  "mlflow==2.20.3",
  "pandas==2.3.2",
  "plotly==5.22.0",
  "psutil==6.0.0",
  "pydantic==2.11.7",
  "pyorb==0.5.4",
  "PyYAML==6.0.2",
  "SciencePlots",
  "scipy",
  "spiceypy==6.0.0",
  "torch==2.2.1",
  "torch_geometric==2.6.1",
  "tqdm",
  "zendir==1.3.1.0",
  "stable-baselines3[extra]",
  "ipython==8.12.3",
  "ipykernel"
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["leo_gym*"]                

[tool.isort]
profile = "black"
line_length = 88
sections = ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
known_first_party = ["leo_gym"]
src_paths = ["src", "tests"]
combine_as_imports = true
import_heading_stdlib = "Standard library"
import_heading_thirdparty = "Third-party"
import_heading_firstparty = "Local"
